# expression: implement vectorized evaluation for `builtinYearWeekWithModeSig`