<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

PCP #12103 

### What problem does this PR solve? <!--add issue link with summary if exists-->

implement vectorized evaluation for `builtinYearWeekWithModeSig`

### What is changed and how it works?

```
$ make vectorized-bench VB_FILE=Time VB_FUNC=builtinYearWeekWithModeSig
cd ./expression && \
                go test -v -benchmem \
                        -bench=BenchmarkVectorizedBuiltinTimeFunc \
                        -run=BenchmarkVectorizedBuiltinTimeFunc \
                        -args "builtinYearWeekWithModeSig"
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-8           1000000000               0.00732 ns/op         0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinYearWeekWithModeSig-VecBuiltinFunc-8             28470             41544 ns/op           29192 B/op          8 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinYearWeekWithModeSig-NonVecBuiltinFunc-8          19538             60255 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      3.505s
```


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test